CPU_Service/requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
google-auth==1.30.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kubernetes==12.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pick==1.0.0
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
psutil==5.8.0
py-healthcheck==1.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
python-dateutil==2.8.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==1.0.1
